requirements.txt
# juntagrico~=1.5.0
git+https://github.com/juntagrico/juntagrico.git@translations
django-modeltranslation
gunicorn==20.0.4
psycopg2==2.8.4
